Add files via upload
"Learning Control in Robotics Trajectory-Based Optimal Control Techniques"

By going through this, it will give us more details of what was in the paper and what are the approaches they used for finding the optimal trajectory path. It is a non-complex way of categorizing things on what I learned during the course work.